q-a-bot-with-django-main/requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
sniffio==1.3.0
sqlparse==0.4.4
